Refactored download-location configs code
